split catalogs module init into files
